refactor: 음식점 리스트 생성 로직 통합 및 필터링 기능 개선
