+++ src/components/layout/tabContainer.js
@@ -1,7 +1,6 @@
-import $restaurantItem from "../restaurant/restaurantItem.js";
+import $createRestaurantList from "../restaurant/restaurantItemList.js";
 import { STORAGE_KEY_NAME } from "../../constants/storage.js";
 import { storageHandler } from "../../utils/storageHandler.js";
-import $restaurantList from "../restaurant/restaurantItemList.js";
 
 const activeTabEvent = (id) => {
   const currentActiveTab = document.querySelector(".select-tab-active");
@@ -12,32 +11,7 @@ const activeTabEvent = (id) => {
 
 const toggleTabClick = (e) => {
   activeTabEvent(e.target.id);
-
-  const restaurantContainer = document.querySelector(
-    ".restaurant-list-container"
-  );
-  restaurantContainer.replaceChildren();
-
-  const categoryFilter = document.getElementById("category-filter").value;
-  const sortFilter = document.getElementById("sorting-filter").value;
-  let restaurantItems;
-  if (e.target.id === "all") {
-    restaurantItems = storageHandler.filterItem(
-      STORAGE_KEY_NAME,
-      categoryFilter,
-      sortFilter
-    );
-  }
-
-  if (e.target.id === "favorite") {
-    restaurantItems = storageHandler.findFavoriteItem(
-      STORAGE_KEY_NAME,
-      categoryFilter,
-      sortFilter
-    );
-  }
-
-  restaurantContainer.appendChild($restaurantList(restaurantItems));
+  $createRestaurantList();
 };
 
 const $tabContainer = (tabs) => {
